requirements.txt
streamlit
ultralytics
opencv-python-headless
pandas
bcrypt
pillows